feat: 페이지 네이밍 수정 및 라우팅(#525)
